[RF][HS3] Absorb RooFitJSONInterface library into RooFitHS3
The separate roofit/jsoninterface package made sense when there was
also a YAML backend, but since only the nlohmann/json backend remains,
it was just unnecessary abstraction boilerplate as a standalone
library.

Move the RooFit::Detail::JSONInterface header into RooFitHS3 and merge
the implementation (JSONInterface.cxx, JSONParser.h, JSONParser.cxx)
into a single translation unit, with the TJSONTree implementation class
now in an anonymous namespace. This remains the only translation unit
that includes nlohmann/json.hpp, so the JSON engine could still be
swapped out in the future by changing only this one file.

Also remove the unused writeYML() interface method, a leftover from the
YAML backend, and move testJSONInterface into the RooFitHS3 test suite.

(cherry picked from commit 240324718dc9b66268041329d4895462f3fecb81)
