+++ roofit/CMakeLists.txt
@@ -19,7 +19,6 @@ if(mathmore)
 endif()
 add_subdirectory(roostats)
 add_subdirectory(histfactory)
-add_subdirectory(jsoninterface)
 add_subdirectory(hs3)
 if(roofit_legacy_eval_backend AND NOT MSVC)
   add_subdirectory(xroofit)
