--- roofit/jsoninterface/LinkDef.h
@@ -1,7 +1,0 @@
-#ifdef __CLING__
-
-#pragma link off all globals;
-#pragma link off all classes;
-#pragma link off all functions;
-
-#endif
