--- roofit/jsoninterface/test/CMakeLists.txt
@@ -1,1 +1,0 @@
-ROOT_ADD_GTEST(testJSONInterface testJSONInterface.cxx LIBRARIES RooFitJSONInterface Matrix)
